+++ build.gradle
@@ -69,6 +69,40 @@ allprojects {
 
 	dependencies {
 		modImplementation "net.fabricmc:fabric-loader:${project.fabric_loader_version}"
+
+		minecraft "com.mojang:minecraft:${project.minecraft_version}"
+		mappings "net.fabricmc:yarn:${project.yarn_mappings}:v2"
+
+		// Required Fabric API modules
+		modImplementation fabricApi.module('fabric-api-base', project.fabric_api_version)
+		modImplementation fabricApi.module('fabric-lifecycle-events-v1', project.fabric_api_version)
+		modImplementation fabricApi.module('fabric-key-binding-api-v1', project.fabric_api_version)
+
+		// Amecs API
+		//modImplementation "de.siphalor:amecsapi-1.19:${project.amecs_api_version}"
+
+		// ModMenu
+		modImplementation "com.terraformersmc:modmenu:${project.modmenu_version}"
+		// Cloth Config
+		modImplementation("me.shedaniel.cloth:cloth-config-fabric:${project.cloth_config_version}") {
+			exclude(group: 'net.fabricmc.fabric-api')
+		}
+
+		if (project.use_third_party_mods == 'true') {
+			configurations {
+				modRuntimeOnly {
+					exclude module: 'fabric-lifecycle-events-v1'
+					exclude module: 'fabric-key-binding-api-v1'
+				}
+			}
+
+			modRuntimeOnly "net.fabricmc.fabric-api:fabric-api:${project.fabric_api_version}"
+			//modRuntimeOnly "maven.modrinth:sodium:${project.sodium_version}"
+			//modRuntimeOnly "maven.modrinth:lithium:${project.lithium_version}"
+			modRuntimeOnly "maven.modrinth:starlight:${project.starlight_version}"
+			//modRuntimeOnly "maven.modrinth:smoothboot-fabric:${project.smoothboot_version}"
+			modRuntimeOnly "curse.maven:no-fade-452768:${project.no_fade_version}"
+		}
 	}
 
 
@@ -97,19 +131,6 @@ allprojects {
 		// with development. Uncomment this line to generate them.
 		// withJavadocJar()
 	}
-}
-
-
-dependencies {
-	include project(path: ':toomanybinds-core', configuration: 'namedElements')
-	include project(path: ':toomanybinds-1.16-compat', configuration: 'namedElements')
-	include project(path: ':toomanybinds-1.17-compat', configuration: 'namedElements')
-	include project(path: ':toomanybinds-1.18.2-compat', configuration: 'namedElements')
-	include project(path: ':toomanybinds-1.19-compat', configuration: 'namedElements')
-	include project(path: ':toomanybinds-1.19.3-compat', configuration: 'namedElements')
-
-	minecraft "com.mojang:minecraft:${minecraft_version_1_19_3}"
-	mappings "net.fabricmc:yarn:${project.yarn_mappings_1_19_3}:v2"
 }
