+++ gradle.properties
@@ -13,91 +13,17 @@ mod_version = 0.3.3
 maven_group = com.github.reviversmc
 archives_base_name = toomanybinds
 
-# Global Dependencies
 fabric_loader_version = 0.14.14
 
-# Module Dependencies
-# 1.16
-minecraft_version_1_16 = 1.16.5
-yarn_mappings_1_16 = 1.16.5+build.10
-fabric_api_version_1_16 = 0.42.0+1.16
-amecs_api_version_1_16 = 1.3.7+mc1.16.5
-modmenu_version_1_16 = 1.16.23
-cloth_config_version_1_16 = 4.15.82
-# 1.17
-minecraft_version_1_17 = 1.17.1
-yarn_mappings_1_17 = 1.17.1+build.65
-fabric_api_version_1_17 = 0.46.1+1.17
-amecs_api_version_1_17 = 1.3.7+mc1.17.1
-nmuk_version_1_17 = 1.0.1+mc1.17
-modmenu_version_1_17 = 2.0.17
-cloth_config_version_1_17 = 5.3.63
-# 1.18.2
-minecraft_version_1_18_2 = 1.18.2
-yarn_mappings_1_18_2 = 1.18.2+build.3
-fabric_api_version_1_18_2 = 0.56.1+1.18.2
-amecs_api_version_1_18_2 = 1.3.6+mc22w06a
-modmenu_version_1_18_2 = 3.2.5
-cloth_config_version_1_18_2 = 6.3.81
-# 1.19
-minecraft_version_1_19 = 1.19.2
-yarn_mappings_1_19 = 1.19.2+build.28
-fabric_api_version_1_19 = 0.70.0+1.19.2
-amecs_api_version_1_19 = 1.3.7+mc22w17a
-nmuk_version_1_19 = 1.0.2+mc1.19.2
-modmenu_version_1_19 = 4.1.2
-cloth_config_version_1_19 = 8.2.88
-# 1.19.3
-minecraft_version_1_19_3 = 1.19.3
-yarn_mappings_1_19_3 = 1.19.3+build.5
-fabric_api_version_1_19_3 = 0.73.2+1.19.3
-amecs_api_version_1_19_3 = 1.3.7+mc1.19.3
-nmuk_version_1_19_3 = 1.0.2+mc1.19.3
-modmenu_version_1_19_3 = 5.0.2
-cloth_config_version_1_19_3 = 9.0.94
+minecraft_version=1.19.4-rc2
+yarn_mappings = 1.19.4-rc2+build.1
+fabric_api_version = 0.75.3+1.19.4
+amecs_api_version= 1.3.7+mc1.19.3
+modmenu_version = 6.1.0-beta.3
+cloth_config_version = 10.0.95
 
 # If true, third-party mods will be loaded during runtime in the developer run configurations
 use_third_party_mods = true
-# 1.16
-lazydfu_version_1_16 = 0.1.2
-sodium_version_1_16 = mc1.16.5-0.2.0
-lithium_version_1_16 = mc1.16.5-0.6.6
-phosphor_version_1_16 = mc1.16.2-0.8.0
-ferrite_core_version_1_16 = 2.1.1-fabric
-smoothboot_version_1_16 = 1.16.5-1.6.0
-no_fade_version_1_16 = 3228523
-# 1.17
-lazydfu_version_1_17 = 0.1.2
-sodium_version_1_17 = mc1.17.1-0.3.4
-joml_version_sodium_1_17 = 1.10.2
-lithium_version_1_17 = mc1.17.1-0.7.5
-starlight_version_1_17 = Starlight 1.0.0 1.17.x
-ferrite_core_version_1_17 = 3.1.0
-smoothboot_version_1_17 = 1.16.5-1.6.0
-no_fade_version_1_17 = 3355800
-# 1.18.2
-lazydfu_version_1_18_2 = 0.1.2
-sodium_version_1_18_2 = mc1.18.2-0.4.1
-joml_version_sodium_1_18_2 = 1.10.2
-lithium_version_1_18_2 = mc1.18.2-0.10.3
-starlight_version_1_18_2 = 1.0.2+1.18.2
-ferrite_core_version_1_18_2 = 4.2.1-fabric
-smoothboot_version_1_18_2 = 3692365
-no_fade_version_1_18_2 = 3550935
-# 1.19
-lazydfu_version_1_19 = 0.1.3
-sodium_version_1_19 = mc1.19.2-0.4.4
-joml_version_sodium_1_19 = 1.10.4
-lithium_version_1_19 = mc1.19.2-0.10.4
-starlight_version_1_19 = 1.1.1+1.19
-ferrite_core_version_1_19 = 5.0.3-fabric
-smoothboot_version_1_19 = 1.19-1.7.1
-no_fade_version_1_19 = 3550935
-# 1.19.3
-lazydfu_version_1_19_3 = 0.1.3
-sodium_version_1_19_3 = mc1.19.3-0.4.8
-lithium_version_1_19_3 = mc1.19.3-0.10.4
-starlight_version_1_19_3 = 1.1.1+1.19
-ferrite_core_version_1_19_3 = 5.1.0-fabric
-smoothboot_version_1_19_3 = 1.19-1.7.1
-no_fade_version_1_19_3 = 3550935
+
+starlight_version = 1.1.1+1.19
+no_fade_version = 3550935
