+++ src/main/java/com/github/reviversmc/toomanybinds/ModConfig.java
