+++ src/main/java/com/github/reviversmc/toomanybinds/integrations/AmecsIntegration.java
@@ -1,0 +1,12 @@
+package com.github.reviversmc.toomanybinds.integrations;
+
+public class AmecsIntegration {
+	/*
+	public static void priorityPressIfApplicable(KeyBinding binding) {
+		if (binding instanceof PriorityKeyBinding) {
+			((PriorityKeyBinding) binding).onPressedPriority();
+		}
+	}
+
+	 */
+}
