+++ src/main/resources/toomanybinds.mixins.json
