--- toomanybinds-1.16-compat/build.gradle
@@ -1,51 +1,0 @@
-sourceCompatibility = JavaVersion.VERSION_1_8
-targetCompatibility = JavaVersion.VERSION_1_8
-
-
-archivesBaseName += '-compat-1.16'
-
-
-dependencies {
-	api project(path: ':toomanybinds-core', configuration: 'namedElements')
-
-	minecraft "com.mojang:minecraft:${project.minecraft_version_1_16}"
-	mappings "net.fabricmc:yarn:${project.yarn_mappings_1_16}:v2"
-
-	// Required Fabric API modules
-	modImplementation fabricApi.module('fabric-api-base', project.fabric_api_version_1_16)
-	modImplementation fabricApi.module('fabric-lifecycle-events-v1', project.fabric_api_version_1_16)
-	modImplementation fabricApi.module('fabric-key-binding-api-v1', project.fabric_api_version_1_16)
-
-	// Amecs API
-	modImplementation "de.siphalor:amecsapi-1.16:${project.amecs_api_version_1_16}"
-
-	// ModMenu
-	modImplementation "com.terraformersmc:modmenu:${project.modmenu_version_1_16}"
-	// Cloth Config
-	modImplementation("me.shedaniel.cloth:cloth-config-fabric:${project.cloth_config_version_1_16}") {
-		exclude(group: 'net.fabricmc.fabric-api')
-	}
-
-
-	if (project.use_third_party_mods == 'true') {
-		configurations {
-			modRuntimeOnly {
-				exclude module: 'fabric-lifecycle-events-v1'
-				exclude module: 'fabric-key-binding-api-v1'
-			}
-		}
-
-		modRuntimeOnly "net.fabricmc.fabric-api:fabric-api:${project.fabric_api_version_1_16}"
-		modRuntimeOnly "maven.modrinth:lazydfu:${project.lazydfu_version_1_16}"
-		modRuntimeOnly "maven.modrinth:sodium:${project.sodium_version_1_16}"
-		modRuntimeOnly "maven.modrinth:lithium:${project.lithium_version_1_16}"
-		modRuntimeOnly "maven.modrinth:phosphor:${project.phosphor_version_1_16}"
-		modRuntimeOnly "maven.modrinth:smoothboot-fabric:${project.smoothboot_version_1_16}"
-		modRuntimeOnly "curse.maven:no-fade-452768:${project.no_fade_version_1_16}"
-	}
-}
-
-
-tasks.withType(JavaCompile).configureEach {
-	it.options.release = 8
-}
