--- toomanybinds-1.16-compat/src/main/java/com/github/reviversmc/toomanybinds/EntrypointRegistrar116.java
@@ -1,10 +1,0 @@
-package com.github.reviversmc.toomanybinds;
-
-public class EntrypointRegistrar116 extends EntrypointRegistrarBase {
-	@Override
-	public void onInitialize() {
-		if (isCompatible("1.16", "1.16.5")) {
-			EntrypointLauncher.setVersionSpecificInitializer(new TooManyBinds116());
-		}
-	}
-}
