--- toomanybinds-1.16-compat/src/main/java/com/github/reviversmc/toomanybinds/TooManyBinds116.java
@@ -1,18 +1,0 @@
-package com.github.reviversmc.toomanybinds;
-
-import net.minecraft.client.MinecraftClient;
-
-import com.github.reviversmc.toomanybinds.autocompletion.VanillaKeybindSuggestions;
-import com.github.reviversmc.toomanybinds.autocompletion.VanillaKeybindSuggestions116;
-
-public class TooManyBinds116 extends TooManyBinds {
-	@Override
-	protected VanillaKeybindSuggestions newVanillaKeyBindSuggestions() {
-		return new VanillaKeybindSuggestions116();
-	}
-
-	@Override
-	protected void openNewLauncherScreen(MinecraftClient client) {
-		client.openScreen(new LauncherScreen116());
-	}
-}
