--- toomanybinds-1.16-compat/src/main/java/com/github/reviversmc/toomanybinds/autocompletion/VanillaKeybindSuggestions116.java
@@ -1,16 +1,0 @@
-package com.github.reviversmc.toomanybinds.autocompletion;
-
-import net.minecraft.client.option.GameOptions;
-import net.minecraft.client.option.KeyBinding;
-
-public class VanillaKeybindSuggestions116 extends VanillaKeybindSuggestions {
-	@Override
-	protected BindSuggestion newBindSuggestion(KeyBinding bind) {
-		return new BindSuggestion116(bind);
-	}
-
-	@Override
-	protected KeyBinding[] allKeys(GameOptions options) {
-		return options.keysAll;
-	}
-}
