--- toomanybinds-1.16-compat/src/main/resources/fabric.mod.json
@@ -1,23 +1,0 @@
-{
-	"schemaVersion": 1,
-	"id": "toomanybinds-compat-1-16",
-	"version": "${version}",
-	"name": "Too Many Binds 1.16 Compat",
-	"environment": "client",
-	"entrypoints": {
-		"main": [
-			"com.github.reviversmc.toomanybinds.EntrypointRegistrar116"
-		]
-	},
-	"depends": {
-		"toomanybinds-core": "*"
-	},
-	"custom": {
-		"modmenu": {
-			"badges": ["library"],
-			"parent": {
-				"id": "toomanybinds"
-			}
-		}
-	}
-}
