--- toomanybinds-1.17-compat/src/main/java/com/github/reviversmc/toomanybinds/EntrypointRegistrar117.java
@@ -1,10 +1,0 @@
-package com.github.reviversmc.toomanybinds;
-
-public class EntrypointRegistrar117 extends EntrypointRegistrarBase {
-	@Override
-	public void onInitialize() {
-		if (isCompatible("1.17", "1.18.1")) {
-			EntrypointLauncher.setVersionSpecificInitializer(new TooManyBinds117());
-		}
-	}
-}
