--- toomanybinds-1.17-compat/src/main/java/com/github/reviversmc/toomanybinds/autocompletion/VanillaKeybindSuggestions117.java
@@ -1,10 +1,0 @@
-package com.github.reviversmc.toomanybinds.autocompletion;
-
-import net.minecraft.client.option.KeyBinding;
-
-public class VanillaKeybindSuggestions117 extends VanillaKeybindSuggestions116 {
-	@Override
-	protected BindSuggestion newBindSuggestion(KeyBinding bind) {
-		return new BindSuggestion117(bind);
-	}
-}
