--- toomanybinds-1.17-compat/src/main/resources/fabric.mod.json
@@ -1,23 +1,0 @@
-{
-	"schemaVersion": 1,
-	"id": "toomanybinds-compat-1-17",
-	"version": "${version}",
-	"name": "Too Many Binds 1.17 Compat",
-	"environment": "client",
-	"entrypoints": {
-		"main": [
-			"com.github.reviversmc.toomanybinds.EntrypointRegistrar117"
-		]
-	},
-	"depends": {
-		"toomanybinds-compat-1-16": "*"
-	},
-	"custom": {
-		"modmenu": {
-			"badges": ["library"],
-			"parent": {
-				"id": "toomanybinds"
-			}
-		}
-	}
-}
