--- toomanybinds-1.18.2-compat/src/main/java/com/github/reviversmc/toomanybinds/EntrypointRegistrar1182.java
@@ -1,10 +1,0 @@
-package com.github.reviversmc.toomanybinds;
-
-public class EntrypointRegistrar1182 extends EntrypointRegistrarBase {
-	@Override
-	public void onInitialize() {
-		if (isCompatible("1.18.2", "1.18.2")) {
-			EntrypointLauncher.setVersionSpecificInitializer(new TooManyBinds1182());
-		}
-	}
-}
