--- toomanybinds-1.18.2-compat/src/main/java/com/github/reviversmc/toomanybinds/TooManyBinds1182.java
@@ -1,11 +1,0 @@
-package com.github.reviversmc.toomanybinds;
-
-import com.github.reviversmc.toomanybinds.autocompletion.VanillaKeybindSuggestions;
-import com.github.reviversmc.toomanybinds.autocompletion.VanillaKeybindSuggestions1182;
-
-public class TooManyBinds1182 extends TooManyBinds117 {
-	@Override
-	protected VanillaKeybindSuggestions newVanillaKeyBindSuggestions() {
-		return new VanillaKeybindSuggestions1182();
-	}
-}
