--- toomanybinds-1.18.2-compat/src/main/resources/fabric.mod.json
@@ -1,23 +1,0 @@
-{
-	"schemaVersion": 1,
-	"id": "toomanybinds-compat-1-18-2",
-	"version": "${version}",
-	"name": "Too Many Binds 1.18.2 Compat",
-	"environment": "client",
-	"entrypoints": {
-		"main": [
-			"com.github.reviversmc.toomanybinds.EntrypointRegistrar1182"
-		]
-	},
-	"depends": {
-		"toomanybinds-compat-1-17": "*"
-	},
-	"custom": {
-		"modmenu": {
-			"badges": ["library"],
-			"parent": {
-				"id": "toomanybinds"
-			}
-		}
-	}
-}
