--- toomanybinds-1.19-compat/build.gradle
@@ -1,55 +1,0 @@
-sourceCompatibility = JavaVersion.VERSION_17
-targetCompatibility = JavaVersion.VERSION_17
-
-
-archivesBaseName += '-compat-1.19'
-
-
-dependencies {
-	api project(path: ':toomanybinds-1.18.2-compat', configuration: 'namedElements')
-
-	minecraft "com.mojang:minecraft:${project.minecraft_version_1_19}"
-	mappings "net.fabricmc:yarn:${project.yarn_mappings_1_19}:v2"
-
-	// Required Fabric API modules
-	modImplementation fabricApi.module('fabric-api-base', project.fabric_api_version_1_19)
-	modImplementation fabricApi.module('fabric-lifecycle-events-v1', project.fabric_api_version_1_19)
-	modImplementation fabricApi.module('fabric-key-binding-api-v1', project.fabric_api_version_1_19)
-
-	// Amecs API
-	modImplementation "de.siphalor:amecsapi-1.19:${project.amecs_api_version_1_19}"
-	// Nmuk version that is compatible with latest Fabric API
-	modImplementation "de.siphalor:nmuk-1.19:${project.nmuk_version_1_19}"
-
-	// ModMenu
-	modImplementation "com.terraformersmc:modmenu:${project.modmenu_version_1_19}"
-	// Cloth Config
-	modImplementation("me.shedaniel.cloth:cloth-config-fabric:${project.cloth_config_version_1_19}") {
-		exclude(group: 'net.fabricmc.fabric-api')
-	}
-
-
-	if (project.use_third_party_mods == 'true') {
-		configurations {
-			modRuntimeOnly {
-				exclude module: 'fabric-lifecycle-events-v1'
-				exclude module: 'fabric-key-binding-api-v1'
-			}
-		}
-
-		modRuntimeOnly "net.fabricmc.fabric-api:fabric-api:${project.fabric_api_version_1_19}"
-		modRuntimeOnly "maven.modrinth:lazydfu:${project.lazydfu_version_1_19}"
-		modRuntimeOnly "maven.modrinth:sodium:${project.sodium_version_1_19}"
-		runtimeOnly "org.joml:joml:${project.joml_version_sodium_1_19}"
-		modRuntimeOnly "maven.modrinth:lithium:${project.lithium_version_1_19}"
-		modRuntimeOnly "maven.modrinth:starlight:${project.starlight_version_1_19}"
-		modRuntimeOnly "maven.modrinth:smoothboot-fabric:${project.smoothboot_version_1_19}"
-		modRuntimeOnly "curse.maven:no-fade-452768:${project.no_fade_version_1_19}"
-	}
-}
-
-
-tasks.withType(JavaCompile).configureEach {
-	// Minecraft 1.18-pre2 upwards uses Java 17.
-	it.options.release = 17
-}
