--- toomanybinds-1.19-compat/src/main/java/com/github/reviversmc/toomanybinds/EntrypointRegistrar119.java
@@ -1,10 +1,0 @@
-package com.github.reviversmc.toomanybinds;
-
-public class EntrypointRegistrar119 extends EntrypointRegistrarBase {
-	@Override
-	public void onInitialize() {
-		if (isCompatible("1.19", "1.19.2")) {
-			EntrypointLauncher.setVersionSpecificInitializer(new TooManyBinds119());
-		}
-	}
-}
