--- toomanybinds-1.19-compat/src/main/java/com/github/reviversmc/toomanybinds/TooManyBinds119.java
@@ -1,11 +1,0 @@
-package com.github.reviversmc.toomanybinds;
-
-import com.github.reviversmc.toomanybinds.autocompletion.VanillaKeybindSuggestions;
-import com.github.reviversmc.toomanybinds.autocompletion.VanillaKeybindSuggestions119;
-
-public class TooManyBinds119 extends TooManyBinds1182 {
-	@Override
-	protected VanillaKeybindSuggestions newVanillaKeyBindSuggestions() {
-		return new VanillaKeybindSuggestions119();
-	}
-}
