--- toomanybinds-1.19-compat/src/main/java/com/github/reviversmc/toomanybinds/autocompletion/VanillaKeybindSuggestions119.java
@@ -1,10 +1,0 @@
-package com.github.reviversmc.toomanybinds.autocompletion;
-
-import net.minecraft.client.option.KeyBinding;
-
-public class VanillaKeybindSuggestions119 extends VanillaKeybindSuggestions1182 {
-	@Override
-	protected BindSuggestion newBindSuggestion(KeyBinding bind) {
-		return new BindSuggestion119(bind);
-	}
-}
