--- toomanybinds-1.19-compat/src/main/resources/fabric.mod.json
@@ -1,23 +1,0 @@
-{
-	"schemaVersion": 1,
-	"id": "toomanybinds-compat-1-19",
-	"version": "${version}",
-	"name": "Too Many Binds 1.19 Compat",
-	"environment": "client",
-	"entrypoints": {
-		"main": [
-			"com.github.reviversmc.toomanybinds.EntrypointRegistrar119"
-		]
-	},
-	"depends": {
-		"toomanybinds-compat-1-18-2": "*"
-	},
-	"custom": {
-		"modmenu": {
-			"badges": ["library"],
-			"parent": {
-				"id": "toomanybinds"
-			}
-		}
-	}
-}
