--- toomanybinds-1.19.3-compat/src/main/java/com/github/reviversmc/toomanybinds/LauncherScreen1193.java
@@ -1,8 +1,0 @@
-package com.github.reviversmc.toomanybinds;
-
-public class LauncherScreen1193 extends LauncherScreen117 {
-	@Override
-	protected void setRepeatKeyboardEvents(boolean repeatEvents) {
-		// 1.19.3 handles this automatically
-	}
-}
