--- toomanybinds-core/build.gradle
@@ -1,31 +1,0 @@
-sourceCompatibility = JavaVersion.VERSION_1_8
-targetCompatibility = JavaVersion.VERSION_1_8
-
-
-archivesBaseName += '-core'
-
-
-dependencies {
-	minecraft "com.mojang:minecraft:${project.minecraft_version_1_16}"
-	mappings "net.fabricmc:yarn:${project.yarn_mappings_1_16}:v2"
-
-	// Required Fabric API modules
-	modImplementation fabricApi.module('fabric-api-base', project.fabric_api_version_1_16)
-	modImplementation fabricApi.module('fabric-lifecycle-events-v1', project.fabric_api_version_1_16)
-	modImplementation fabricApi.module('fabric-key-binding-api-v1', project.fabric_api_version_1_16)
-
-	// Amecs API
-	modImplementation "de.siphalor:amecsapi-1.16:${project.amecs_api_version_1_16}"
-
-	// ModMenu
-	modImplementation "com.terraformersmc:modmenu:${project.modmenu_version_1_16}"
-	// Cloth Config
-	modImplementation("me.shedaniel.cloth:cloth-config-fabric:${project.cloth_config_version_1_16}") {
-		exclude(group: 'net.fabricmc.fabric-api')
-	}
-}
-
-
-tasks.withType(JavaCompile).configureEach {
-	it.options.release = 8
-}
