--- toomanybinds-core/src/main/java/com/github/reviversmc/toomanybinds/integrations/AmecsIntegration.java
@@ -1,12 +1,0 @@
-package com.github.reviversmc.toomanybinds.integrations;
-
-import de.siphalor.amecs.api.PriorityKeyBinding;
-import net.minecraft.client.option.KeyBinding;
-
-public class AmecsIntegration {
-	public static void priorityPressIfApplicable(KeyBinding binding) {
-		if (binding instanceof PriorityKeyBinding) {
-			((PriorityKeyBinding) binding).onPressedPriority();
-		}
-	}
-}
