--- toomanybinds-core/src/main/resources/fabric.mod.json
@@ -1,32 +1,0 @@
-{
-	"schemaVersion": 1,
-	"id": "toomanybinds-core",
-	"version": "${version}",
-	"name": "Too Many Binds Core",
-	"environment": "client",
-	"entrypoints": {
-		"main": [
-			"com.github.reviversmc.toomanybinds.EntrypointLauncher"
-		]
-	},
-	"mixins": [
-		"toomanybinds.mixins.json"
-	],
-	"depends": {
-		"minecraft": ">=1.16",
-		"fabric-lifecycle-events-v1": "*",
-		"fabric-key-binding-api-v1": "*",
-		"cloth-config2": ">=4.12.41"
-	},
-	"recommends": {
-		"amecs": "*"
-	},
-	"custom": {
-		"modmenu": {
-			"badges": ["library"],
-			"parent": {
-				"id": "toomanybinds"
-			}
-		}
-	}
-}
